improve :: Removes the hack in JomSocial
Fields are added in JomSocial itself, so we do not need to hack JomSocial code.
We also have 2 triggers to manipulate the fields, so no need to hack profile model.
JS has changed response of different ajaxified tasks, so we need to work on that.
Temporary reposnse is change to make it working, but need to work.

ref #689 <changelog>